+++ ChangeLog.md
@@ -19,6 +19,7 @@ QIIME 1.7.0-dev (changes since QIIME 1.7.0 go here)
 * Renamed check_id_map.py to validate_mapping_file.py for clarity.
 * Increased default rdp_max_memory from 1500M to 4000M as this was almost always needing to be increased when re-training on modern reference databases.
 * Required biom-format version is now 1.3.1.
+* Required matplotlib version is now >= 1.1.0, <= 1.3.1.
 
 QIIME 1.7.0 (14 May 2013)
 =========================